Fix lint issue: Rename Component
