+++ databrowser/src/components/tabs/TabItem.vue
